Refactor project layout and add shopping functionality
The layout of the project was restructured to a classic shop page style. A ShopItemClass component with the ability to showcase individual items in the shop was added. This was done to fulfil user requirements for a more shopping oriented web application, rather than the initial boilerplate design. AppVeyor was also set up for continuous integration.

In addition, 'page.module.css' was removed as this file was replaced by 'globals.css' to centralize and manage all global styles. Updated '.gitignore' file to include '/.idea' hence preventing IDEA config files from being tracked.