Update background image paths in globals.css
The paths for the background images in globals.css were updated from their root level paths "/item-blue.jpg" and "/item-black.jpg", respectively to "ra-components-store-class/item-blue.jpg" and "ra-components-store-class/item-black.jpg". This change was necessary due to the restructuring of the project directory, in which the images were moved into the "ra-components-store-class" directory.